# ARROW-9131: [C++] Faster ascii_lower and ascii_upper.