@maartenbreddels 2020-06-15
I also thought that we could do a bit check instead of the range check, e.g. `code_unit & 0b11100000) == 0b01100000`, but that would also transform the backtick for instance (binary value 0b1100000).

The generated code looks vectorized indeed. I didn't look into the details of the generated code by clang and GCC, it seems their performance is a bit different, so we might be able to squeeze out a bit more if we want. Happy to look into that later (create a new issue), but I rather spend my time on other functions now.