@pitrou 2020-06-15
It's ok, there's no need to further optimize those functions.