@pitrou 2020-06-15
Travis-CI failure is unrelated.