@maartenbreddels 2020-06-15
Thanks, let me know if my workflow is ok, or if I can make things go smoother.

PS: I am looking for a document describing the kernel design. I see these two cases (`if (batch[0].kind() == Datum::ARRAY) {` and the else clause, but I am not sure I fully understand this. But I'm not sure where this is described, if it is.

PS2: Are there alternative channels for quick/small questions, or is this fine?