@pitrou 2020-06-15
I think the current approach (implement kernels one-by-one) is reasonable and manageable for us (and for you as well I hope).

I don't think there's much documentation for now around the kernel design. Basically, kernels should usually be able to process two kinds of inputs (represented as `Datum`s): arrays and scalars. That said, arrays are the dominant case, so if we leave scalars unimplemented in a given kernel, that's not an urgent problem.

For quick development questions, we have a public chat instance at https://ursalabs.zulipchat.com/ - just register though and you can chat with the team. The main channel there is the "dev" channel, and Zulip allows you to create subtopics - don't hesitate to use those!