@maartenbreddels 2020-06-15
> I think the current approach (implement kernels one-by-one) is reasonable and manageable for us (and for you as well I hope).

No, this is fine.



> That said, arrays are the dominant case, so if we leave scalars unimplemented in a given kernel, that's not an urgent problem.

👍