@wesm 2020-06-15
> PS2: Are there alternative channels for quick/small questions, or is this fine?

Can we use dev@arrow.apache.org so everyone can see the discussion and it's searchable via Google later?