@maartenbreddels 2020-06-15
> Can we use [dev@arrow.apache.org](mailto:dev@arrow.apache.org) so everyone can see the discussion and it's searchable via Google later?

I'm ok with that if not considered too noisy (like small cmake questions etc).