@wesm 2020-06-15
It shouldn't be a problem.